rename move-pair to pair and fix versions and submodune names
