+++ submodules/pair/collect.go
@@ -1,4 +1,4 @@
-package move_pair
+package pair
 
 import (
 	"context"
@@ -17,7 +17,7 @@ import (
 	ibctm "github.com/cosmos/ibc-go/v8/modules/light-clients/07-tendermint"
 
 	opchildtypes "github.com/initia-labs/OPinit/x/opchild/types"
-	"github.com/initia-labs/kvindexer/submodules/move-pair/types"
+	"github.com/initia-labs/kvindexer/submodules/pair/types"
 )
 
 const (
