+++ submodules/pair/submodule.go
@@ -1,4 +1,4 @@
-package move_pair
+package pair
 
 import (
 	"context"
@@ -17,7 +17,7 @@ import (
 
 	"github.com/initia-labs/kvindexer/collection"
 	pairtypes "github.com/initia-labs/kvindexer/pair/types"
-	"github.com/initia-labs/kvindexer/submodules/move-pair/types"
+	"github.com/initia-labs/kvindexer/submodules/pair/types"
 	kvindexer "github.com/initia-labs/kvindexer/x/kvindexer/types"
 )
