+++ submodules/wasm-nft/types/keys.go
@@ -5,7 +5,7 @@ const (
 	SubmoduleName = "wasm-nft"
 
 	// Version is the current version of the submodule
-	Version = "v0.1.0"
+	Version = "v0.1.2"
 )
 
 // store prefixes
